Move duplicate methods into AbstractTemplateTool
`RadiusTemplateTool` and `RadiusCellTemplateTool` are the only direct children of `AbstractTemplateTool`. They both had
identical `#getPen()` and `#getPenForOverlay()` methods, so these have been merged into `AbstractTemplateTool`.
